requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
pydantic-settings
aider-chat
pytest
